requirements_gpu.txt
numpy
onnx
onnxruntime-gpu
cdsapi
netCDF4
os
ncvue